timezonechange/addeventlistener-timezonechange-fired.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>

function defaultTimeZone() {
  return (new Intl.DateTimeFormat()).resolvedOptions().timeZone;
}
promise_test(async t => {
  const oldTimeZone = "Pacific/Fakaofo";
  const newTimeZone = "Asia/Taipei";
  // First, setup the timezone in a fixed place.
  await window.test_driver.set_timezone(oldTimeZone);
  assert_equals(defaultTimeZone(), oldTimeZone);
  return new Promise(r => {
    window.addEventListener('timezonechange', r);
    window.test_driver.set_timezone(newTimeZone);
  }).then(e => {
      assert_equals(defaultTimeZone(), newTimeZone);
  });
}, "Test that the timezonechange event fires on window.addEventListener('timezonechange')");
</script>
</body>
</html>
